refactor: split conductor.py into smaller focused modules
Reorganize conductor.py into a modular structure for better maintainability:

- Created conductor/ subdirectory with dedicated modules
- Extracted configuration management to config.py
- Moved Netbox query utilities to netbox_utils.py
- Separated Ironic synchronization logic into ironic_sync.py
- Extracted helper functions (deep_compare, deep_merge, etc.) to utils.py
- Moved Celery app initialization to app.py
- Centralized task definitions in tasks.py
- Updated main conductor.py to re-export all public interfaces

AI-assisted: Claude Code
