+++ osism/tasks/conductor/__init__.py
@@ -1,0 +1,54 @@
+# SPDX-License-Identifier: Apache-2.0
+
+import copy
+from celery import Celery
+from celery.signals import worker_process_init
+from loguru import logger
+
+from osism.tasks import Config
+from osism.tasks.conductor.config import get_configuration
+from osism.tasks.conductor.ironic import sync_ironic as _sync_ironic
+
+
+# App configuration
+app = Celery("conductor")
+app.config_from_object(Config)
+
+
+@worker_process_init.connect
+def celery_init_worker(**kwargs):
+    pass
+
+
+@app.on_after_configure.connect
+def setup_periodic_tasks(sender, **kwargs):
+    pass
+
+
+# Tasks
+@app.task(bind=True, name="osism.tasks.conductor.get_ironic_parameters")
+def get_ironic_parameters(self):
+    configuration = get_configuration()
+    if "ironic_parameters" in configuration:
+        # NOTE: Do not pass by reference, everybody gets their own copy to work with
+        return copy.deepcopy(configuration["ironic_parameters"])
+
+    return {}
+
+
+@app.task(bind=True, name="osism.tasks.conductor.sync_netbox")
+def sync_netbox(self, force_update=False):
+    logger.info("Not implemented")
+
+
+@app.task(bind=True, name="osism.tasks.conductor.sync_ironic")
+def sync_ironic(self, force_update=False):
+    _sync_ironic(get_ironic_parameters, force_update)
+
+
+__all__ = [
+    "app",
+    "get_ironic_parameters",
+    "sync_netbox",
+    "sync_ironic",
+]
